refactor(agents): 将头像解析逻辑集中到 utils，消除多层 prop drilling
